new start
I'm programming in vscode because pycharm was giving me issues. for some reason though I keep getting the error pkg_resources aren't found. Idk if it is the current version of brainflow that is causing issues because it doesn't look like pkg_resources exists anymore. I don't know how to go back to an older version of brainflow though