Working test
I got the TestConnection.py program working. It can read data from the EEG through the dongle. File location: C:\Users\kotag\Desktop\EEG_software\TestConnection. The thing that is weird is that you have to right click on the file and click on “Run Python File in Terminal”. By doing this the program runs and doesn’t run in the pkg_resources error. I also had to uninstall and reinstall python 5.12.0 and vscode. I also fixed my file system as well. I am not going to use pycharm, maybe later but it was giving me issues that I don’t care to fix right now. I need to think of a way so the wires stay put and the clips don’t break. That goes the same with the electrodes, I don’t want to buy more electrodes.