PythonGUI V1
I'm just starting the project with a basic screen, testing button options, drop down menu, and so on before we implement the actual EEG stuff. I am using pygame for this but I might change it base on more research but right now it seems to be the easiest to use.